view: support for adding highlight annotations
The interaction in ev-view can be summarized as:
button press -> create annotation
motion notify -> save annotation (update
quadrilaterals)
button release -> save annotation (save bounding
box in the annotation mapping)

The bounding box on the third step is obtained via
poppler (poppler_annot_get_rectangle), but I am
not sure this rectangle is updated as the
annotation grows with more quadrilaterals... I am
afraid not.
The responsibility of updating the quadrilaterals
and saving the bounding box to the mapping are all
of the back-end, which in turn uses poppler
functions. Ev-view's only task is to compute a doc
rectangle to pass as an argument to the back-end
functions.